feat/test: 完善循环，条件语句翻译，实现多维数组结构, 添加测试用例

+++ include/parser.h
@@ -20,7 +20,7 @@ class TokenParser {
 private:
     LineNo LineInfo;    
     double DoubleNumVal;
-    long IntNumVal;
+    long long IntNumVal;
     std::string LiteralVal;
     std::string IdStr;
     int LastChar;
@@ -36,7 +36,7 @@ public:
 
     LineNo getCurLineNo() const { return LineInfo; }
     double getDoubleVal() const { return DoubleNumVal; }
-    long   getIntVal()    const { return IntNumVal; }
+    long long getIntVal()    const { return IntNumVal; }
     const std::string& getIdStr() const { return IdStr; }
     const std::string& getLiteral() const { return LiteralVal; }
     bool   isSigned() { return IsSigned; }
